Move MessageParser into its own new file-set, seperate from MessageUtils file-set
Create MessageParser.cpp and .h and relocate the MessageParser class into these files,
Make necessary changes to reference new location of MessageParser where needed.
Make changes to test sources and non-solution files that need MessageParser.
