+++ dds/DCPS/security/CryptoBuiltInImpl.cpp
@@ -21,6 +21,7 @@
 
 #include "dds/DCPS/RTPS/MessageUtils.h"
 #include "dds/DCPS/RTPS/MessageTypes.h"
+#include "dds/DCPS/RTPS/MessageParser.h"
 #include "dds/DCPS/RTPS/RtpsCoreTypeSupportImpl.h"
 
 #include <openssl/err.h>
